+++ src/pages/home/MoguriModal.css
@@ -1,0 +1,151 @@
+@import url('https://fonts.googleapis.com/css2?family=Nanum+Pen+Script&display=swap');
+
+.modal-overlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 1000;
+}
+
+.modal-content-home {
+  background: #F8E28E;
+  padding: 20px;
+  border-radius: 10px;
+  border: solid 1px black;
+  position: relative;
+  width: 300px;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
+  font-family: "Nanum Pen Script", cursive;
+  font-size: 18px;
+  color: black;
+  bottom: 20px;
+  right: 250px;
+}
+
+.modal-header {
+  margin-bottom: 10px;
+  color: #557C56;
+}
+
+.modal-close-button {
+  position: absolute;
+  top: 10px;
+  right: 10px;
+  background: none;
+  border: none;
+  font-size: 20px;
+  cursor: pointer;
+}
+
+.modal-close-button img {
+  width: 20px; 
+  height: 20px; 
+}
+
+.modal-back-button {
+  position: absolute;
+  top: 10px;
+  left: 210px;
+  background: none;
+  border: none;
+  cursor: pointer;
+}
+
+.modal-back-button img {
+  width: 20px; 
+  height: 20px; 
+}
+
+.modal-input {
+  background-color: white;
+  padding: 10px;
+  margin-left: 10px;
+  border: 1px solid black;
+  border-radius: 5px;
+  margin-bottom: 10px;
+  font-family: "Nanum Pen Script", cursive;
+  font-size: 18px;
+  color: black;
+}
+
+.modal-button {
+  padding: 10px;
+  background-color: #FFFFFF;
+  color: black;
+  border: 1px solid black;
+  border-radius: 5px;
+  cursor: pointer;
+  font-family: "Nanum Pen Script", cursive;
+  font-size: 18px;
+  display: block;  
+  margin-top: 10px; 
+  margin-left: 217px;
+}
+
+.modal-button:hover {
+  background-color: #D9D9D9;
+}
+
+.modal-close-button:hover {
+  background-color: #F8E28E;
+}
+
+.modal-back-button:hover {
+  background-color: #F8E28E;
+}
+
+.modal-result {
+  font-family: "Nanum Pen Script", cursive;
+  font-size: 18px;
+  margin-top: 10px;
+}
+
+.evolved-modal-content {
+  background: #C7F9FC;
+  padding: 20px;
+  border-radius: 15px;
+  border: solid 1px black;
+  position: relative;
+  width: 350px;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
+  font-family: "Nanum Pen Script", cursive;
+  font-size: 20px;
+  color: black;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  filter: opacity(95%);
+}
+
+.selected-moguri-evolved {
+  height: 150px;
+  margin-bottom: 20px;
+}
+
+.evolved-modal-header {
+  margin-bottom: 20px;
+  color: #557C56;
+  font-size: 24px;
+}
+
+.evolved-modal-button {
+  padding: 5px 15px;
+  background-color: #D9D9D9;
+  color: black;
+  border: 1px solid black;
+  border-radius: 5px;
+  cursor: pointer;
+  font-family: "Nanum Pen Script", cursive;
+  font-size: 18px;
+  margin-top: 20px;
+}
+
+.evolved-modal-button:hover {
+  background-color: #BFBFBF;
+}
